Fixed an issue with precompiled headers
precompiled headers would not be sorted, and duplicates were not removed.
Because of this, if a file included a header that was already used in the pch, it would be included again, requiring the header to be built again.
This also happened if includes were rearranged in the source.

The includes and defines used are now sorted and have no duplicates.